fix : 미션 api 오류 수정 - 미션 진행도 float으로 받아오면서 생기는 문제 수정

+++ data/src/main/kotlin/jinproject/stepwalk/data/repositoryImpl/MissionRepositoryImpl.kt
@@ -143,7 +143,7 @@ class MissionRepositoryImpl @Inject constructor(
                             id = 0,
                             designation = missionResponse.designation,
                             type = type,
-                            achieved = if (localAchieved >= detail.currentValue) localAchieved else detail.currentValue,
+                            achieved = if (localAchieved >= detail.currentValue) localAchieved else detail.currentValue.toInt(),
                             goal = detail.goal
                         )
                     )
